Gerrit https:// prefix processed at a central location
There are a few places that do the https:// normalization, some of them append and some of them trim, it's not always easy to tell whether https:// should be expected. Consolidating them to avoid accidents.

This change also adds a IsGerritOrg method by checking the prefix of https:// or http://, this is not something new, as this is used only in places where this assumption is implicitly applied.