Tide for Gerrit: comment on batch merges
In case of flaky tests, Tide triggered prowjobs for highest priority PR might fail even when batch prowjobs passed. And in this case Crier would report this failure on the PR before Tide merges the PR, this might cause confusing to users so comment on the PR explaining that the merge was based on batch testing.