kettle: combine failure message and backtrace
Ginkgo v2 started splitting the information about a failure into the XML
attribute (failure message) and data (stack backtrace). Kettle used to pick
only the data, with the result that go.k8s.io/triage showed that.

It's debatable whether the backtrace is relevant for triaging. It was used in
the past and it might help to distinguish failures with the same message but a
different origin, so now both are stored again.