Change behavior of setting job queue concurrency to 0
Previously setting this value to 0 meant that there is no limit on
how many jobs may be executed concurrently. This does not make much
sense as this should be achieved by simply not configuring the job
queue.

Instead, the new behavior will prevent any job from being executed.
This is especially useful when temporarily stopping workloads when
no resources are available.

The old behavior can be still triggered by setting the concurrency
to a negative value.