ci: Update COS version for containerd jobs
Bump containerd jobs image family to use cos-stable. cos-stable
currently points to COS M101.

I suggest we point image-family to use cos-stable, so we will always use
the latest stable version of COS so that the version does not get stale
and we don't need to manually update the job config.

Signed-off-by: David Porter <porterdavid@google.com>