test: Add node e2e cgroupv1 job
We currently have parity between cgroupv1 and cgroupv1 jobs for the
following 3 jobs:

1. cos-cgroupv1-containerd-node-e2e / cos-cgroupv2-containerd-e2e
2. cos-cgroupv1-containerd-e2e / cos-cgroupv2-containerd-e2e
3. cos-cgroupv1-containerd-node-features / cos-cgroupv2-containerd-node-features

However we are missing a cgroupv1 variant of the serial job.

cgroupv2 there is a serial job (cos-cgroupv2-containerd-node-e2e-serial)
but no direct corresponding cgroupv1 job.

We should add a corresponding job so there is direct parity between
cgroupv2 and cgroupv1 containerd jobs.

Signed-off-by: David Porter <porterdavid@google.com>