Add jobs for the `stable4` version marker
This commit updates `test_config.yaml` with jobs for the `stable4`
version marker. This commit is currently NOOP because we already have
those jobs and we already generated jobs for 1.26, but this ensures that
our scripts will NOT delete `stable4` jobs when generating the new jobs
(e.g. for 1.27).

The added `stable4` jobs were copied from the `stable3` jobs. I ran the
job generation locally and there's no diff for the `stable4` jobs in
terms of commands, args, and env. We can also always adjust the added
jobs later if needed.

Signed-off-by: Marko Mudrinić <mudrinic.mare@gmail.com>