config-forker: perform replacement on the command
config-forker is currently performing replacement only on the job's
args. This is causing an issue where ci-kubernetes-build job is not
forked properly because it uses `command` instead of `args`. Because of
that, forked ci-kubernetes-build jobs are using an incorrect extra
version marker. This is now fixed with this commit by ensuring we
perform replacement on both command and args.

Signed-off-by: Marko Mudrinić <mudrinic.mare@gmail.com>