config-rotator: update version markers before running strings.Replace
We are doing two things when updating version markers in config-rotator:
- run strings.Replace to replace all occurences of the old marker with
  the new marker
- bump the old marker to the next one by calling
  `updateGenericVersionMarker`

This brings an issue that we're bumping the version markers for two
instead of for one. For example, we bump k8s-stable1 to k8s-stable3.

That's because we first run strings.Replace("k8s-stable1", "k8s-stable2")
and then we run `updateGenericVersionMarker` function which also updates
the version marker.

This has never been an issue before because config-forker wasn't
properly generating jobs, but that was fixed with a previous commit.

Signed-off-by: Marko Mudrinić <mudrinic.mare@gmail.com>