add granular_approval support for approve plugin This plugin is an updated approve plugin that supports granular approvals and regex in the OWNERS files
This commit adds a new PR status section to the approval notifier message.
This section shows the status of approvals of the folders involved in the PR.
It shows if all the files under a folder are completely approved, partially approved
or unapproved. The folders used are the ones where the owners files assocaited
with the original PR reside.

This commit also refactors to approvel plugin code to
have 2 different code paths when using simple vs granular approvals.

This commit also introduces a new `approveHelper` interface
that is used to handle the different approval mechanisms.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>
Co-authored-by: Yuvaraj Kakaraparthi <kakaraparthy@vmware.com>