replace GitReferenceDynamic and GitReferenceStatic with prowcrd.Refs
The new gangway.Refs struct matches up 1:1 with the existing
prowcrd.Refs type. The changes in here make the compiler happy, but is
not complete because we will now make Sub use our
CreateJobExecutionRequest type. Once that happens, we can revisit
Gangway and make it use the same underlying logic that Sub uses for
converting a CreateJobExecutionRequest into a Prow Job CR.