Add a webhook job reporter
This patch adds a 'webhook' reporter_config for Prow jobs.

---

The configuration accepts a URL where to POST the JSON-serialised
ProwJob information. For future-proofness, the `v1.ProwJob` object is
wrapped in a JSON object that contains a single `prow_job` property.

---

In order to authenticate the actual webhooks HTTP calls, the calls are
digitally signed with a user-provided Ed25519 key. The signature is sent
in a `x-prow-token` header.

The signed digest is a concatenation of:

1. the target url
2. an RFC3339-formatted timestamp of the current time in the UTC timezone
3. the request body

A single null byte separates the fields.

The same timestamp is included in the request header `x-prow-timestamp`.

Note that this patch does not provide a public key dissemination
mechanism.