approve: Rectify flow of approve2 on approve GH review
Previously, on submitting an approve review on github while
making use of granular approval (for ex: /approve files path),
this would be treated as the highest level of approval for the
user who was approving the PR. For ex: if there are the files:

- x/x.go
- y/y.go
- z/z.go

And user X has approval rights on dirs `x/` and `z`, if X submits
an approve review on GH with the comment `/approve files x/*`, the
previous implementation would consider files x/x.go AND z/z.go as
approved. This is unintuitive and we would lile ONLY x/x.go to be
approved.

This commit rectifies that.
This means that the behaviour of approve2 in terms of calculating
approval status when an approval is given, is the same whether
reviewActsAsApproval is enabled or not. The difference comes in
changes requested review.

This commit also adds additional tests for this behaviour.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>