approve: Refactor approveHelper interface
Refactor the approveHelper interface to re-use as many
methods as possible and reduce boilerplate.

In order to do this, some methods of the original approve now
also take a `path` argument which is ignored. This is done to
have common function signatues across approve and approve2 in
order to abstract away implementations into a common interface.

This commit also makes other changes nescessary to support this
refactor.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>