Capture CCM logs from tests.
We are migrating to cloud provider code running in the CCM.
It should no longer be running in the KCM, KAS or Kubelet.
Several of our tests now run the CCM as a result.
We are also trying to make more of the tests run the CCM.
However our current test runs do not capture the logs of the CCM.