spyglass: add warning for negative time durations
Sometimes, naughty (out of band) processes can tamper with the timestamp
fields in finished.json. This can result in spyglass showing negative
time durations.

The change here notices such anomalies and surfaces it to the user,
informing them of the naughtiness.

Because the CI tests run with UTC as the timezone, we refrain from
checking the entire rendered string (because locally, we could end up
with a different timezone).