Deduplicate CRI-O job templates
We now de-duplicate things which are required by all tests, like
installing the tools, setting up SELinux as well as installing CRI-O.
This way we achieve easier maintainability without risking to forget
anything if we want to add new jobs.

This PR is no actual logical change to the config, just changing the
order of the systemd units within the ignition file.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>