add InRepoConfigCache metrics
We can now measure how often keys get deleted (evicted), forcefully (by
the LRU caching library we use) or manually (by our application code).

If the rate of evictions is rather high, this means that the cache size
is too small.  This should help make deployment configurations easier to
reason about so that we can detect scale issues ahead of time.

We also track the total number of lookups, hits, and misses for the
cache.

All of these metrics are built into the InRepoConfigCache, and are
passed as callbacks into the lower-level LRUCache. This is so that the
LRUCache can remain as a generic cache for future use in other places.
This also enables us to add unit tests (directly on the LRUCache)
without having to worry about mocking out the Prometheus metrics that
come built-in with the InRepoConfigCache.