experiment: add kind-logs-e2e-k8s.sh
This is a verbatim copy of sigs.k8s.io/kind/hack/ci/e2e-k8s.sh, commit
e9188fd867. The script needs to be modified to increase kubelet's
containerLogMaxSize settings, something that the original kind script doesn't
support and will not support because it is too specific. Users who need to
customize the behavior of that script are meant to fork it.