node: ci: Enable `fork-per-release` for jobs running NodeSpecialFeature
To ensure that serial jobs with node specific special features are carried over
to 1.28 and onwards we add add annotation to enable `fork-per-release`.

It is worth noting that that for these jobs `optional` is set to true and
`always_run` is set to `false` making these jobs optionally available if
needed for testing (backport) PRs made to the release branches.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>