needs-rebase: plugin test: fix data race race
Instead of using locks, set up the test so that the "sleep" global
variable is pointing to the correct function (in this case a NOP
function). This is simpler and avoids the race that still exists
here (namely in the deferred function where we still write to the global
variable).