trigger github action workflows from prow
currently if a github workflow needs to be retriggered in containerd repo,
only the org members can do it. But after having an ok-to-test label on
the PR, the author also should be able to retrigger the workflows so that
flaky jobs and jobs which have the changes can be easily iterated on.

Signed-off-by: Akhil Mohan <makhil@vmware.com>