integration testing: configure for testing with race detection
When reducing parallelism (KUBE_INTEGRATION_TEST_MAX_CONCURRENCY) and extending
the allowed runtime (KUBE_TIMEOUT), integration testing works with race
detection enabled.

This was tested in https://github.com/kubernetes/kubernetes/pull/116980 with
locally modified settings. The job definitions are the right place for settings
which depend on the capability of the test environment.