fix: fit cli-utils stress test on k8s-infra
The cli-utils stress test was recently switched back to the k8s-infra
prow cluster. This cluster has smaller nodes available compared to the
EKS cluster, which was resulting in this job being unschedulable. Other
jobs using this cluster request up to 7 CPU, so this should fit on a
node in the k8s-infra cluster.