e2e-node-canary: use image config to remount /tmp
This job uses '--images=cos-97-16919-29-21' option, which
uses cos image as is. /tmp is mounted with 'noexec' option in
COS, which makes it impossible to run ginkgo and other test binaries.

Usage of 'image-config' option ensures remounting /tmp with 'exec'
option. This is done in the jobs/e2e_node/containerd/init.yaml:
runcmd:
  - echo "Test run from /tmp folder, remounting it"
  - mount /tmp /tmp -o remount,exec,suid