webhook-server: pin to a known-working base image
With ko v0.11.2, the base image for webhook-server (if unspecified)
defaults to

   gcr.io/distroless/static:nonroot@sha256:9ecc53c269509f63c69a266168e4a687c7eb8c0cfd753bd8bfcaa4f58a90876f

but with ko v0.14.1, it defaults to

   cgr.dev/chainguard/static:latest@sha256:67a1b00e0134e2b3a614c7198a26f7deed9d11b7acad4d52c79c0cfd47a2eae7

and breaks the app, because it can no longer write to paths in the
container's filesystem.

Fix the breakage by pinning to the previous default version, so that the
new chainguard default does not get selected any more.