Provide HeadRef for tide github pull requests
When tide runs a serial build of a presubmit job, it is better if the PULL_HEAD_REF environment variable can be set for the job (as when run by trigger).  This change preserves the HeadRef data so that it can be propagated to the job in this case so the presubmit build environment is more consistent in tide.