integration test: refresh git config from existing primary clone
The minimal fetches feature from 02308e0beb (Merge pull request #30453
from listx/minimal-fetch, 2023-08-22), was reverted due to errors seen
in production. It was observed that our integration tests failed to
catch the condition of refreshing an existing primary clone. This is
because

(1) we only check inrepoconfigs in our integration tests for sub, and

(2) all of our tests in the sub component (which tests using
inrepoconfigs against a real remote git server (fakegitserver)) were set
up to only process requests for newly-seen repos.

This means that we never exercised the "reuse an existing primary mirror
clone" code path that 02308e0beb changed (and which was buggy), and is
why the tests passed there.

Add a new test to exercise the above code path. It was tested locally
with a build of the sub component at 02308e0beb, and this test case
failed there, as expected.

Also, populate the fakegitserver before running the test cases. This way
we don't have to run the tests serially. That is, instead of
inrepoconfig-presubmit3-repeat relying on inrepoconfig-presubmit3 always
running first (and making the primary mirror clone being created),
either one can run ahead of the other --- as long as both test cases are
exercised, we'll hit the code path where one of them results in the
prowYAMLGetter seeing that the repo has already been cloned (and where
we only need to just refresh the existing clone).