git v2 client: force creation of a new variable
It has been observed that sometimes we end up passing in the same SHA
object into FetchCommits. As the baseSHA can never be the same as a SHA
in the headSHAs, this is a bug with our code.

Specifically, we used to use a inrepoconfigRepoOpts global variable, and
one of its fields was a string slice (FetchCommits). This meant that we
could have been assigning to the same slice field multiple times from
multiple threads.

Avoid the global variable, and instead use a function to return a new
value each time.