Revert "git v2 client: keep the primary clone up to date"
This reverts commit e46df0a581d22545f3c77e3106594ab328f71f2a.

We've thought of ways to make this faster. The main issue with this is
that we block secondary clones from happening while the primary fetches
the base SHA. Also, for presubmits, we still have to do a second fetch
of the headSHA anyway, so this technically slows things down.

Revert for now while we explore better options.