CAPK: require some checks
cluster-api-provider-kubevirt repo uses several github actions. But
these actions (e.g. unit test, lint and so on) are not gate for PR
merges. This fact cause external contributors' PRs to sometimes get
merged with no verification.

This PR requires the unit tests, the linter, the check-gen and the
e2e-integration action to pass before merging a PR into the main branch.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>