prow: extend RepoOwner functionality to return owners aliases
RepoOwner by itself could not be used to retrieve owner aliases even though
the information was present in the cache entries.

Extending RepoOwner into a new interface allows us to not break existing
users of the repoowners client as much as possible.

The commit also makes changes to prow plugins that use this interface.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>