go-deps.sh: modernize option processing
We preserve the previous convention of using "--minor" and "--patch"
flags to edit the behavior of `go get`, as well as the ability to
specify specific packages as trailing arguments. That is, the previous
invocation of "$0 --minor|--patch [packages...]" is still supported in
this change. Running the script without any arguments also still behaves
the same way (it just runs 'go mod tidy' in the root and hack/tools
folders).

Also:
- Break up logical units of work into their own separate functions.
- Add new "--tools" flag to do go dependency maintenance in the
  hack/tools folder.
- Add new "--only-tidy" flag to only run "go mod tidy" in both the root
  and hack/tools folders.
- Add usage guide with examples of how to use the script.
- Use STDERR (>&2) for printing informational text.