sub: compute Gerrit-style ref earlier
There is no benefit of delaying this computation until we hit
the (overloaded) getProwJobSpec() method. Instead, compute it when we
run peToCjer() when we first create the JobExecutionRequest.

This change is a preparatory step for the introduction of a new
interface, InRepoConfigGetter, in subsequent commits (to make that
transition easier).