moonraker: promote client timeout to central config
Instead of hardcoding the client-side timeout into each binary that
needs to use Moonraker, put it into the main config. This way we don't
need to restart the services to pick up changes to this setting (it will
be auto-mounted already into the Config via the Config Agent, which
listens to the ConfigMap for the main config).