integration test: add env var duplication test case
The added test in this change will fail with an error like the
following:

    {
    "branch": "master",
    "component": "sub",
    "error": "job trigger-inrepoconfig-postsubmit-via-pubsub-repo7 failed to merge presets for podspec: env var duplicated in pod spec: FOO_FILE",
    "file": "k8s.io/test-infra/prow/gangway/gangway.go:761",
    "func": "k8s.io/test-infra/prow/gangway.(*postsubmitJobHandler).getProwJobSpec",
    "level": "info",
    "msg": "Failed to get postsubmits from inrepoconfig",
    "org": "https://fakegitserver.default/repo/org1",
    "orgRepo": "https://fakegitserver.default/repo/org1/repo7",
    "repo": "repo7",
    "severity": "info",
    "time": "2023-09-23T07:04:27Z"
    }

And this is because we currently do defaulting of the retrieved
inrepoconfig jobs on both the server side (within moonraker) and on the
client side (in the moonraker client). The defaulting of the client side
was added in cc4560f75f (Merge pull request #30823 from
airbornepony/moonraker-defaulting-fix, 2023-09-22) to address a separate
problem.

In the next commit, we'll make defaulting *not* happen on the server
side (so that not only do we avoid adding env vars twice, but also avoid
other non-idempotent effects lurking in our defaulting logic). This is
the correct approach because it is a waste to do any defaulting on the
server side if the client ultimately still has to default it themselves
on their end.