tide: use richer logging
Just a few lines above, we already create the more verbose logger with
all of the details regarding the refs. We forgot to reuse this
logger (which we already do in other areas in this function), so start
using it here too.

The discrepancy was detected when we started using moonraker for tide in
Google's internal Prow instance, when after the switchover, tide
stopped logging the refs for pulls that had merge conflicts (because
previously those conflicts were reported by the InRepoConfigCache, and
because moonraker is the one using the cache, those errors were only
reported inside moonraker's logs and not tide's logs).