node: resource-mgrs: Run resource manager tests on multi-NUMA
Currently in `image-config-serial-resource-managers.yaml` we have
config with images that point to GCP node with multi-numa support
(n2d-standard-32) as well as without it (n1-standard-4).

This results in the jobs that refer to this config to be executed
twice, once on a node with multi-numa support and another time
without it which explain the flaky topology manager tests and
tests being skipped when they are executed on nodes that
don't provide this support with the following message:

```
 [SKIPPED] this test is intended to be run on a multi-node NUMA system
```

Let's ensure that we use ubuntu and cos based images such that they both
have multi-numa numa support for execution of resource manager tests.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>