kubernetes-e2e-gce-cos-alpha-features: enable KUBE_WATCHLIST_INCONSISTENCY_DETECTOR
When this environmental variable is set then
the reflector performs a data consistency check.

The consistency check is meant to be enforced only in the CI, not in production.
The check ensures that data retrieved by the watch-list api call
is exactly the same as data received by the standard list api call.