[sig-testing] Add `ci-kubernetes-e2e-kind-rootless`
The job runs kubetest2 with rootless kind: https://kind.sigs.k8s.io/docs/user/rootless/

kubetest2 is executed with "kindinv" (Kubernetes in (Rootless) Docker in (GCE) VM) driver:
https://github.com/rootless-containers/kubetest2-kindinv

GCE VM (Ubuntu 22.04) is used for creating a host with cgroup v2 and systemd.
Using KRTE was not an option, as it does not support systemd (discussed in PR 30744).

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>