kubeadm-presubmits: remove EKS canary jobs
These jobs do not work. They start and after ~N seconds
they return a Pod error. To be able to merge PRs in k/kubeadm
we must pass "/skip" everytime.

We can switch the existing jobs to "eks-prow-build-cluster"
once it's clear what's going on. Maybe it's due to the
"tolerations" and "nodSelector" sections.