inrepoconfig: add integration test to check for advancing the base branch
This adds a new integration test case for Moonraker. It first makes
Moonraker retrieve the repo with master pointing to a particular commit.
Then it makes a Moonraker client call with the "Refs.BaseRef" populated
to a non-empty string, with the expectation that Moonraker will advance
the branch on its end (because Refs.BaseRef is now not empty). Most of
the code is based on the existing TestMoonrakerBurst test case.