shouldSkipProcessingChange: refactor to use positive booleans for readability
Previously this function was written as shouldSkipProcessingChange() and
we had to negate it in order to use it.

Remove the extra level of boolean indirection by flipping the return
values. This aligns with the child call to triggerJobs() (which is also
a positive action), making the code easier to read. Also rename to
`shouldTriggerJobs` for simplicity. Update test case descriptions to
use shorter, simpler language.