ResultStore: determine content type from filenames
Retreiving content type from GCS requires per-file GCS access via the
opener.Attributes function which is expensive, and untenable when many
files are involved.

While we believe ResultStore and/or its UI should take care of fetching
file attributes, that is not the case today. So to resolve this TODO, we
will take the length from the io.Iterator and determine the type using
the mime package.

An override is present for *.log files, to use text/plain instead of
text/x-log.