pull-kubernetes-e2e-gce: enable KUBE_WATCHLIST_INCONSISTENCY_DETECTOR
When this environmental variable is set then
the reflector performs a data consistency check.

The consistency check is meant to be enforced only in the CI, not in production.
The check ensures that data retrieved by the watch-list api call
is exactly the same as data received by the standard list api call.

As of today the inconsistency detector runs with pull-kubernetes-e2e-gce-cos-alpha-features
where it is used only by a signle test.

As we are preparing to promote the WatchList feature to beta and enable it for KCM,
this job will be used to check for any inconsistencies.