config-bootstrapper: delete keys that were not updated (untouched)
The config-bootstrapper works by looking at local files, and treating
all files it finds as a file addition. Then these files are converted
into "ConfigMapUpdate" entries, and sent to updateconfig.Update() to
update the corresponding ConfigMap key.

The problem is that the above mechanism doesn't work for deleting keys.
For example, deleting a file and then running config-bootstrapper again
will simply "update" all other files except the deleted one (because it
doesn't see it in the local filesystem).

Teach updateconfig.Update() how to delete ConfigMap keys that are not
updated. Do this by calling updateconfig.MarkStaleKeysForDeletion()
which checks for those ConfigMap keys that were left untouched and
creates corresponding ConfigMapUpdate entries signaling their deletion.

An alternative way would be to call the "patch" verb to delete those
keys individually. However, this would incur additional API overhead of
needing to call both "update" (to send the locally modified ConfigMap
object) and the "patch" verb (to delete untouched keys), not to mention
that we would be bypassing the existing mechanism we already have in
updateconfig.Update(), which already knows how to delete keys inside the
locally modified ConfigMap object [1]).

For these reasons, continue to use the "Update" verb without "Patch". We
have to send the entire object over (instead of only sending a diff),
but we've already been doing this and we don't want to make large
changes in this sensitive area. The previous commit records for the
record why using "Update" isn't very accurate in terms of unit tests,
but this is a minor cost to pay in exchange for making this commit much
simpler (no need to modify how deletions are done at [1]).

[1] https://github.com/kubernetes/test-infra/blob/d3ef1505f12cd2dd9f2bfc898bb987dc22c48066/prow/plugins/updateconfig/updateconfig.go#L140-L144