config-bootstrapper: document limitations in ConfigUpdater
The config-bootstrapper currently is unable to delete keys in the
ConfigMap, because it simply treats all files in the sourcePaths as key
"additions". Therefore it only knows about key additions, not deletions.

Underneath, the config-bootstrapper calls the run() function [1] which
takes a ConfigUpdater object [2] as a parameter. Ultimately, if it sees
that it needs to update a ConfigMap, it invokes updateconfig.Update [3],
which uses the "update" verb for modifying the ConfigMap in the K8s API
server [4]. Presumably, we could also use the "patch" verb as the K8s
client-go package supports this method for the ConfigMapInterface [5].
The differences between Update and Patch are described (not very well)
in the upstream docs [6].

The point though is that we use the Update verb. And the fake Kubernetes
client library implements this Update verb by replacing the entire
object wholesale (after creating it) [7]. This is why the test we add in
this commit passes, even though it should fail (underneath we end up
creating a brand new object from scratch instead of actually modifying
the underlying object). Indeed, the library warns that the
implementation is overly simplified and doesn't cover all use cases of
Update [8].

We add this test case nevertheless to capture the current behavior in
more detail. The next commit will implement key deletions using the
higher-level ConfigMapUpdate struct in updateconfig.go.

[1] https://github.com/kubernetes/test-infra/blob/9b88591d62f9d775a8c2611e96b6b5e2389b1cee/prow/cmd/config-bootstrapper/main.go#L106
[2] https://github.com/kubernetes/test-infra/blob/d3ef1505f12cd2dd9f2bfc898bb987dc22c48066/prow/plugins/config.go#L544
[3] https://github.com/kubernetes/test-infra/blob/9b88591d62f9d775a8c2611e96b6b5e2389b1cee/prow/cmd/config-bootstrapper/main.go#L153
[4] https://github.com/kubernetes/test-infra/blob/d3ef1505f12cd2dd9f2bfc898bb987dc22c48066/prow/plugins/updateconfig/updateconfig.go#L184
[5] https://pkg.go.dev/k8s.io/client-go@v0.29.1/kubernetes/typed/core/v1#ConfigMapInterface
[6] https://kubernetes.io/docs/reference/using-api/api-concepts/#patch-and-apply
[7] https://github.com/kubernetes/client-go/blob/b0ed45df66138ec7761c6db922d8f96672773dd5/testing/fixture.go#L411
[8] https://github.com/kubernetes/client-go/blob/b0ed45df66138ec7761c6db922d8f96672773dd5/kubernetes/fake/clientset_generated.go#L132-L136