pin alpine image to latest (specific) tag
Some nodes are multiple-months old and will use a version of
gcr.io/k8s-prow/alpine that is rather old now (although at the time that
the image was pulled, it was "new" because `:latest` was pointing to the
newest version at that time).

Pin the image to specific version. This is simpler for auditing.

This will also force older nodes to upgrade to the newer version pinned
here to pick up vulnerability fixes we've incorporated into the newer
alpine versions recently.