Increase max rate limit for crier controller rate limiter
Previously we were using the workqueue.DefaultControllerRateLimiter(),
which is calling workqueue.NewMaxOfRateLimiter internally with
rate.Limit set to 10 and bucket size set to 100. And the setting for
NewItemExponentialFailureRateLimiter is 5ms for base delay and 1000
seconds for max delay.

In this change, we double the QPS and bucket size and decrease max delay
by half to 500 seconds and keep the base delay intact.