ResultStore: do not rely on returned Invocation
When the ResultStore writer creates an invocation, it recorded it in
the retInv field. It was subsequently used to determine the invocation
name. This use was overlooked when adding the capability to resume
uploads, since it retInv will be nil when resuming an upload.

This change eliminates retInv and stores the invID field instead.