kubernetes_e2e.py: remove working dir check if `--build` is specified
The job `pull-release-cluster-up` is failing right now with:

```
+ /workspace/scenarios/kubernetes_e2e.py --build=quick --cluster= --down=false --extract=local --gcp-node-image=gci --gcp-nodes=4 --gcp-zone=us-west1-b --provider=gce --test_args=--ginkgo.focus=definitely-not-a-real-focus --timeout=65m
Traceback (most recent call last):
  File "/workspace/scenarios/kubernetes_e2e.py", line 391, in <module>
    main(parse_args())
  File "/workspace/scenarios/kubernetes_e2e.py", line 259, in main
    raise ValueError(k8s)
ValueError: /home/prow/go/src/github.com/kubernetes/release
```

Reason for that is that the current working director is already
`/home/prow/go/src/github.com/kubernetes/release`, while the script
expects to be in `kubernetes/kubernetes`. It does not seem to matter
much for `kubetest` in which working directory we are, means we can
remove the check completely. Also the method `add_k8s` does not do
anything at all which means we can remove it, too.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>