sig-scalability: add scheduler_perf presubmit
The periodic job has been failing for a while, which only got noticed when the
release team started to check for the 1.31.0 alpha. I am volunteering to check
dashboard notification emails for this because most of the recent changes in
scheduler_perf and its configuration have been made by me.

To help with future changes, let's add an optional, non-blocking presubmit job
that can be run manually to verify such changes before merging them.