sig node: support different API versions in DRA presubmit
This is a first attempt to make the presubmits support a PR which bumps the API
version (like https://github.com/kubernetes/kubernetes/pull/125488).

If this works, the other two DRA jobs also will get converted.
The quoting of the command line flags is a bit more readable
this way.