sig-instrumentation: fix and update ci-kubernetes-kind-e2e-json-logging
The job broke when the DRA API was bumped from v1alpha2 to v1alpha3. To avoid
that in the future and also to get more coverage, now all alpha and beta features
get enabled.

Some other parts get updated while at it, including the switch to
--label-filter. SIG Storage tests don't have labels (but perhaps they should?),
so they are not included anymore. This probably reduces coverage and is worth
revisiting later.