Use main branch from cloudprovider
Cloud provider main branch and cluster api 1.16 know how to build the HPC version of the cloudprovider image. Since we don't use cloud provider for much in upstream e2e we should be ok to use the mis matched versions here.  If it become an issue we can backport the cloud provider changes to older versions

Signed-off-by: James Sturtevant <jsturtevant@gmail.com>