Use new location for Prow images instead of old.
Prow images have been published to a new location for a while now
(us-docker.pkg.dev/k8s-infra-prow/images/). These should be the same
images as published to the old location, so update images to pull from
the new location instead.

Also updated a couple of spots that reference the old gcr.io/k8s-prow
location (in the README.md and the Makefile).