Migrate AWS CI janitor to community infra
Ref:
  - https://github.com/kubernetes/test-infra/issues/33129

`maintenance-ci-aws-janitor` is running against a historical AWS account
currently used by the kOps project for CI purposes. It's unclear if
other projects used this account but we should move it outside of
Google-owned build clusters to unblock the Prow migration.
AFAIK `kops-infra-prow-kops-build` already has the credentials in place to run
it.