Migrate TestGrid locations to gs://kubernetes-ci-logs.
Replace references in TestGrid config to gs://kubernetes-jenkins with
gs://kubernetes-ci-logs (the new bucket) instead.

This includes a change to the generated release jobs. I updated the logs
location to kubernetes-ci-logs and ran:

`./releng/generate_tests.py --yaml-config-path=releng/test_config.yaml`

(This included a change to
config/jobs/kubernetes/generated/generated.yaml, but I am not including
that change in this since it shouldn't be relevant to this update.)