ipvs jobs
Reduce the number of jobs used for ipvs and leave the minimum necessary
to understand the feature works, there is nobody paying attention or
fixing the issues and this adds noise that can mask real issues.

For the jobs that still remain skip the flaky tests, that seems
concentrated around the webhooks e2e feature, once somebody fix them
they can be unskiped.