remove pull-kubernetes-verify-lint
The original intent behind this merge-blocking job was to enforce that certain
linter issues get fixed in new or modified code for "intermediate severity"
issues (not necessary to fix them in existing code, but should not occur in new
code).

This approach had some drawbacks:
- It forced developers touching code which has such issues to address them,
  whether it makes sense in their PR or not.
- Reverting a PR which fixed such an issue as part of some other change
  failed in pull-kubernetes-verify-lint because the revert is treated
  like "updated code" and re-introduces the issue.
- It's not entirely clear which issues fall into this "intermediate severity".

Therefore the job gets removed. These issues still get reported as part of the
optional pull-kubernetes-linter-hints job. Developers and reviewers are
encouraged to pay attention to it and use their own judgement to decide whether
something reported their needs to be addressed in a PR.

Kubernetes itself still has the hack/golangci-strict.yaml configuration. Once
this PR is merged, that can and probably should be simplified.