fake pull-kubernetes-verify-lint
Removing the required pull-kubernetes-verify-lint caused a problem for pending
PRs: tide is now "Waiting for status to be reported", but no such status is
going to appear because the job is gone. This blocks merging those PRs.

Authors would have to close the PR and create a new one. Closing and reopening
is not enough. Both is confusing. A simpler solution is to have a job with
the same name which doesn't do anything and thus can be run to provide the
expected status.