dra: move all jobs into a single file
This makes editing a bit easier because one doesn't have to jump back and forth
between different files while trying to keep periodic and presubmit jobs in
sync. In the past, changes were made to one but not the other, perhaps because
author and reviewer forgot about the other half of the jobs.

It also enables usage of YAML anchors and aliases to define some settings only
in one place. They are scoped to one document.