dra: add canary jobs
Breaking the jobs which are in use while making changes is annoying, but hard
to avoid because testing these jobs locally is difficult. The
https://docs.prow.k8s.io/docs/build-test-update/#how-to-test-a-prowjob method
doesn't work (or at least not easily) because of nested containers (kind inside
kind, for E2E) and the need for a special test environment (E2E node).