dra: add shell script for applying canary changes
The benefit of trying out changes in canary jobs is diminished if the actual
change then still needs to be done manually. There has been at least one case
elsewhere where the canary job changes were okay, but then copying them into
the real jobs was bungled such that they broke.

To avoid this, the shell script automates copying of changes. To use it, run
dra-sync.sh on a new, clean branch and submit the generated commit in a PR.