CAPO: Conformance test without CI artifacts
The conformance suite has been using the latest unreleased version of
kubernetes. However, this is quite unstable and has not been working now
for a few weeks. Since we don't run any other tests against the same
kubernetes version, I suggest we switch the conformance suite to use the
same kubernetes version that we use elsewhere.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>