DRA: promote beta CI jobs to release informing
The intent was first announced in December in
https://groups.google.com/a/kubernetes.io/g/dev/c/76qs9sPiD-o/m/6MnoHeEaAQAJ,
with one agreement and no objections. From that email:

After the promotion of the core Dynamic Resource Allocation (KEP 4381 =
DRA with structured parameters) to beta in 1.32, I'd like to suggest
that we promote the corresponding periodic CI jobs (ci-kind-dra,
ci-node-*-dra) to "release informing".

These jobs are already closely monitored by the DRA developers
(including alerting), but it would be good to increase visibility and
get the release team officially involved.

The ci-kind-dra job covers E2E testing. It has to be a separate job
because the cluster setup must enable DRA. The ci-node-*-dra jobs cover
E2E node testing for key runtimes (containerd, CRI-O with cgroup v1 and
v2). They have been stable leading up to the 1.32 release and we fully
intend to keep it that way.

Links:
- process: https://github.com/kubernetes/sig-release/blob/master/release-blocking-jobs.md#promoting-or-demoting-jobs
- dashboard: https://testgrid.k8s.io/sig-node-dynamic-resource-allocation
- triage:
- https://storage.googleapis.com/k8s-triage/index.html?job=ci-kind-dra
- https://storage.googleapis.com/k8s-triage/index.html?job=ci-node-e2e-cgroupv1-crio-dra
- https://storage.googleapis.com/k8s-triage/index.html?job=ci-node-e2e-cgroupv2-crio-dra
- https://storage.googleapis.com/k8s-triage/index.html?job=ci-node-e2e-containerd-1-7-dra