DRA: reduce period for forked jobs
We run periodic jobs every 6 hours (now specified in the template, for
visibility and because there is no need to override it by job) and should
reduce that when forking for a release. 24 hours is the maximum used for
kind-e2e-parallel. In contrast to that one, we directly reduce the period
instead of using a staggered reduction (e.g. 12 hours for the most recent
release, then 24).