pull-kubernetes-integration: split into Go and cmd testing
It's annoying that pull-kubernetes-integration also runs "make test-cmd"
because it makes the job run longer (which is a problem because it is one of
the slowest, thus delaying merging of PRs) and because of the additional,
confusing output produced by the shell scripts which tests commands.

These two canary jobs test
https://github.com/kubernetes/kubernetes/pull/130325.

If successful, pull-kubernetes-integration and the corresponding periodic
job(s) will be updated to run only "make-integration" (as the name implies!)
and new *-cmd jobs will get added.