pull-kubernetes-integration: split into Go and cmd testing, for real
It's annoying that pull-kubernetes-integration also runs "make test-cmd"
because it makes the job run longer (which is a problem because it is one of
the slowest, thus delaying merging of PRs) and because of the additional,
confusing output produced by the shell scripts which tests commands.

A trial run with canary jobs was successful. However, the test scripts then
got renamed with an additional "test-" prefix. This has been taken account
when modifying the actual jobs.