node: update presubmit jobs make sure to use --skip-regex
The topology manager `-kubetest2` lanes where not running any test (see:
Related to https://github.com/kubernetes/kubernetes/issues/125950)

Turns out that the `kubetest2` node tester, which we need to use
in these lanes, has a `SkipRegex` default setting which include the
`Serial` tag.

Problem: all the e2e_node suites are serial, because they need to mutate
the node, which is a giant shared piece of state.

So, the fix is to always use `--skip-regex` to the values we need

Along the way, take the chance to move to label filtering, which
we wanted to do since long time.

Signed-off-by: Francesco Romani <fromani@redhat.com>